% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daner-package.R
\docType{package}
\name{daner-package}
\alias{daner}
\alias{daner-package}
\title{daner: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
