% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{divipola_provins}
\alias{divipola_provins}
\title{DIVIPOLA province-level codes}
\format{
A tibble with 97 rows and 17 variables:
\describe{
  \item{provincia}{Province name}
  \item{dept_cod}{2-digit DANE department code}
  \item{dept_nom}{Department name, upper case}
  \item{long}{Longitude}
  \item{lat}{Latitude}
  \item{region_nom}{Region name}
  \item{region_cod}{Region code}
  \item{dept_nom_corto}{Department name, title case, shortened for display}
  \item{muni_nom_corto}{Municipality name, title case, shortened for display}
  \item{nunique_dept_cod, nunique_dept_nom, nunique_long, nunique_lat,
    nunique_region_nom, nunique_region_cod, nunique_dept_nom_corto,
    nunique_muni_nom_corto}{Diagnostic counts, from `data-raw/divipola.R`,
    of how many distinct values of each variable exist within a province
    (all should be 1, i.e. these variables are constant within a
    province)}
}
}
\source{
\url{http://geoportal.dane.gov.co:8084/Divipola/}
}
\usage{
divipola_provins
}
\description{
Colombia's official political-administrative division codes (DIVIPOLA)
aggregated at the province ("provincia") level, a grouping of
municipalities used mainly in Cundinamarca and a few other departments.
}
\keyword{datasets}
