% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{fix_dept_cod}
\alias{fix_dept_cod}
\title{Make DANE-code a character and the 2-length format}
\usage{
fix_dept_cod(dept_cod)
}
\arguments{
\item{dept_cod}{dane code for department}
}
\value{
a character vector
}
\description{
Sometimes useful, when you get the code as numeric and for any reason want to
have it in its canonical format (string of length 2)
}
