% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{fix_muni_cod}
\alias{fix_muni_cod}
\title{Make DANE-code a character and the 5-length format}
\usage{
fix_muni_cod(muni_cod)
}
\arguments{
\item{muni_cod}{dane code for munincipality}
}
\value{
a character vector
}
\description{
Sometimes useful, when you get the code as numeric and for any reason want to
have it in its canonical format (string of length 5)
}
