% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{get_ddi_vars}
\alias{get_ddi_vars}
\title{Get DDI Variables}
\usage{
get_ddi_vars(doc_url_txt)
}
\arguments{
\item{doc_url_txt}{A character vector containing a URL to an XML document or
XML text (to be read by `xml2::read_xml`), or an already parsed XML object
(class `xml_document` or `xml_node`).}
}
\value{
A data frame containing information about each variable in the
DDI document. The data frame includes columns for:

* `name`: Variable name (from `name` attribute)
* `id`: Variable ID (from `ID` attribute)
* `files`: File name(s) associated with the variable (from `files` attribute)
* `intrvl`: Interval information (from `intrvl` attribute)
* `labl`: Variable label(s) (extracted from `labl` nodes and cleaned)
* `catgry`: List containing information (catValu, labl) for var categories
}
\description{
This function retrieves information about variables from a DDI
(Data Documentation Initiative) XML document.
It first uses `get_ddi_xml` to parse the XML document and then extracts
information for each variable node using XPath expressions.
}
\examples{
\dontrun{
get_ddi_vars("http://microdatos.dane.gov.co/index.php/catalog/ddi/419")
}

}
