% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{get_ddi_xml}
\alias{get_ddi_xml}
\title{Get a XML document, parsing the text received or the document in the URL}
\usage{
get_ddi_xml(doc_url_txt)
}
\arguments{
\item{doc_url_txt}{A character vector containing a URL or XML text
(to be read by `xml2::read_xml`), or an already parsed XML object
(class `xml_document` or `xml_node`).}
}
\value{
The parsed XML document as an object of class `xml_document`.
}
\description{
This function attempts to retrieve a DDI (Data Documentation Initiative) XML
document from a provided source. It can handle:
}
\details{
* A character vector containing a URL to an XML document.
* A character vector containing a XML document.
* An already parsed XML object of class `xml_document` or `xml_node`.

It is used as a helper function to ensure a valid xml in other functions
This function is memoised (memoise::memoise) to prevent parsing every time
it's called with the same argument (or even worse, downloading every time)
}
\examples{
\dontrun{
get_ddi_xml("http://microdatos.dane.gov.co/index.php/catalog/ddi/419")
}

}
\seealso{
`xml2::read_xml`, `xml2::xml_ns_strip`
}
