% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_zipped_sav.R
\name{ingest_zipped_sav}
\alias{ingest_zipped_sav}
\title{Ingest .sav files from within a .zip file}
\usage{
ingest_zipped_sav(zipfile_path, ...)
}
\arguments{
\item{zipfile_path}{path to the .zip file}

\item{...}{further arguments passed to haven::read_sav}
}
\value{
a tibble with columns
\itemize{
\item sav_file (character): path of the sav file within the zip file
\item size (numeric): size of the uncompressed sav file
\item date (of class "POSIXct"): date of the sav file
\item sav_data (tibble): the data in sav_file, ingested via haven::read_sav
}
}
\description{
This is a common format for microdato DANE.
Why bother?, you could simply call haven::read_sav passing the path to the
zip file and haven is smart enough to decompress it for you.
Well, yeah, in theory. But there are a few glitches that end up breaking
that approach. First, Microdato Dane sometimes bundles several different
files in a single .zip file (e.g. a .sav, along plain-text -csv, txt- and
sometimes also .dta files). This creates trouble for haven
(haven usually can decompress the file and warn you there are several files,
but frequently end up failing). In addition, sometimes the name of the .sav
file does not match that of the .zip file, also creating issues. Finally, in
a few cases, several .sav files are bundled in the .zip file, and this
function allows you to ingest them all at once.
}
\details{
So this function unzip the .sav files contained in the zip
}
