% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{join_divipola_dept_cod}
\alias{join_divipola_dept_cod}
\title{Join DIVIPOLA department variables into lhs}
\usage{
join_divipola_dept_cod(
  lhs,
  by,
  dept_vars = c("dept_cod", "dept_nom", "long", "lat", "dept_nom_corto")
)
}
\arguments{
\item{lhs}{A data.frame-like}

\item{by}{A character vector indicating the column in `lhs` that contains
department code}

\item{dept_vars}{A character vector to customize which variables from
DIVIPOLA to bring into `lhs`}
}
\value{
lhs with additional DIVIPOLA variables
}
\description{
Left join lhs and DIVIPOLA at the department level. Useful to quickly bring
basic data from departments (name, long, lat) to a dataset that contains
only department code.
}
\examples{
df <- data.frame(cod = c("05", "25", "11"))
join_divipola_dept_cod(df, by = "cod")
}
