% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{list_ddi_files}
\alias{list_ddi_files}
\title{List all 'fileDscr' nodes in the ddi_xml}
\usage{
list_ddi_files(doc_url_txt)
}
\arguments{
\item{doc_url_txt}{A character vector containing a URL or XML text
(to be read by `xml2::read_xml`), or an already parsed XML object
(class `xml_document` or `xml_node`).}
}
\value{
data.frame/tibble with two columns: id and uri of the file
}
\description{
DDI XML should include one 'fileDscr' node for each dataset in the study
and other nodes refer to the 'fileDscr' ID to link, for example, variables
to their respective dataset.
This function aims to help you find the ID of the file you are working with,
by listing the 'fileDscr' arguments, that include the URI (with the name at
the end) and the file ID.
}
\examples{
\dontrun{
list_ddi_files("http://microdatos.dane.gov.co/index.php/catalog/ddi/210")
}
}
