% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provins}
\alias{provins}
\title{Municipality-to-province lookup}
\format{
A tibble with 1,123 rows and 5 variables:
\describe{
  \item{nombre_depto}{Department name, upper case}
  \item{provincia}{Province name}
  \item{codigo_municipio}{5-digit DANE municipality code}
  \item{nombre_mpio}{Municipality name, upper case}
  \item{nombre}{Municipality name, title case}
}
}
\source{
\url{https://www.dane.gov.co/files/censo2005/provincias/subregiones.xls}
}
\usage{
provins
}
\description{
A lookup table mapping municipalities to the province ("provincia") they
belong to. Used to build [divipola_centropob] and [divipola_provins].
}
\keyword{datasets}
