% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daner-package.R
\docType{package}
\name{daner-package}
\alias{daner}
\alias{daner-package}
\title{daner: Helpers for Working with DANE Data}
\description{
Helpers for working with data published by DANE (Departamento Administrativo Nacional de Estadistica, Colombia's national statistics agency). Provides functions to standardize and join DIVIPOLA political-administrative division codes, parse DDI (Data Documentation Initiative) XML metadata published alongside DANE's microdata catalogs, and ingest the zipped SPSS files DANE distributes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edalfon/daner}
  \item \url{https://edalfon.github.io/daner/}
  \item Report bugs at \url{https://github.com/edalfon/daner/issues}
}

}
\author{
\strong{Maintainer}: edalfon \email{edalfon@gmail.com}

Authors:
\itemize{
  \item edalfon \email{edalfon@gmail.com}
}

}
\keyword{internal}
