% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{join_divipola_muni_cod}
\alias{join_divipola_muni_cod}
\title{Join DIVIPOLA municipality variables into lhs}
\usage{
join_divipola_muni_cod(
  lhs,
  by,
  muni_vars = c("dept_cod", "centro_poblado_cod", "dept_nom", "muni_nom",
    "centro_poblado_nom", "tipo_centro_poblado", "long", "lat", "distrito", "tipo_muni",
    "area_metropolitana", "region_nom", "region_cod", "dept_nom_corto", "provincia",
    "muni_nom_corto")
)
}
\arguments{
\item{lhs}{A data.frame-like}

\item{by}{A character vector indicating the column in `lhs` that contains
municipality code}

\item{muni_vars}{A character vector to customize which variables from
DIVIPOLA to bring into `lhs`}
}
\value{
lhs with additional DIVIPOLA variables
}
\description{
Left join lhs and DIVIPOLA at the municipality level. Useful to quickly
bring basic data from municipalities (name, department, long, lat) to a
dataset that contains only municipality code.
}
\examples{
df <- data.frame(cod = c("05001", "25001"))
join_divipola_muni_cod(df, by = "cod")
}
